Simplify CGO switch of video converters
Switch on build time and define hasCGOConvert as a const.
